Measure how often each branch is active during an AP
calculate how often a dendritic branch is active simultaneously as a
somatic calcium transient. This is a useful control to check that all
imaged branches were active at least a few times during the experiments,
and that there are no branches that have no calcium channels. Also
compared data acquired during the first half of an experiment with
second half of an experiment, to control for bleaching.